Changing the size type from ushort to long

+++ Assets/com.alephvault.unity.meetgard/Runtime/Types/MessageHeader.cs
@@ -27,7 +27,7 @@ namespace AlephVault.Unity.Meetgard
                 serializer.Serialize(ref MessageSize);
             }
 
-            public void CheckSize(ushort maxMessageSize)
+            public void CheckSize(long maxMessageSize)
             {
                 if (MessageSize > maxMessageSize)
                 {
